BuffSystem Implementation
Bullet damage now calculated by the bullets themselves, buffsystem now works, Glimpse into the Ether implemented

+++ Assets/Systems/DeletionSystem.cs
@@ -19,9 +19,11 @@ public class DeletionSystem : ComponentSystem
         
         Entities.ForEach((Entity e, ref DeleteComp d, ref Translation t) =>
             {
-                
+                if(d.buffer <= 0)
                     World.Active.EntityManager.DestroyEntity(e);
+                d.buffer--;
             });
+
         
     }
 }
